+++ packages/Panel/src/components/Dialog/Dialog.styles.js
@@ -3,9 +3,9 @@ import tokens from "@paprika/tokens";
 import styled, { css, keyframes } from "styled-components";
 import { slideFromDirections } from "../../slideFromDirections";
 
-const childPanelCss = ({ isCompact, groupOffsetY, offsetY }) => {
+const childPanelCss = ({ isCompact, groupOffsetY, offset }) => {
   const totalOffset = isCompact ? 48 : 64;
-  const childBottomOffsetY = totalOffset + groupOffsetY + offsetY;
+  const childBottomOffsetY = totalOffset + groupOffsetY + offset.top;
   return css`
     border-bottom-left-radius: ${tokens.spaceSm};
     border-top-left-radius: ${tokens.spaceSm};
@@ -40,11 +40,63 @@ const compactStyles = css`
   padding: ${stylers.spacer(2)};
 `;
 
+function getSlideOutTransform(slideFrom) {
+  switch (slideFrom) {
+    case slideFromDirections.LEFT:
+      return "translateX(-100%)";
+    case slideFromDirections.RIGHT:
+      return "translateX(100%)";
+    case slideFromDirections.BOTTOM:
+      return "translateY(100%)";
+    default:
+      return null;
+  }
+}
+
+function getSlideInTransform(slideFrom) {
+  switch (slideFrom) {
+    case slideFromDirections.LEFT:
+    case slideFromDirections.RIGHT:
+      return "translateX(0)";
+    case slideFromDirections.BOTTOM:
+      return "translateY(0)";
+    default:
+      return null;
+  }
+}
+
 export const Dialog = styled.div(
-  ({ width, isAnimating, slideFrom, isOpen, hasPushedElement, kind, offsetY, zIndex, isInline }) => {
-    const widthAsPx = Number.isNaN(Number(width)) ? width : `${width}px`;
-    const slideOutTransform = slideFrom === slideFromDirections.LEFT ? "translateX(-100%)" : `translateX(100%)`;
-    const slideInTransform = "translateX(0)";
+  ({ hasPushedElement, height, isAnimating, isInline, isOpen, kind, offset, slideFrom, width, zIndex }) => {
+    const slideOutTransform = getSlideOutTransform(slideFrom);
+    const slideInTransform = getSlideInTransform(slideFrom);
+
+    let _height = null;
+    let _width = null;
+    let left = null;
+    let right = null;
+    let top = null;
+    let bottom = null;
+    let borderTop = null;
+
+    switch (slideFrom) {
+      case slideFromDirections.LEFT:
+      case slideFromDirections.RIGHT:
+        _height = offset.top ? `calc(100% - ${offset.top}px);` : "100%";
+        _width = Number.isNaN(Number(width)) ? width : `${width}px`;
+        left = slideFrom === slideFromDirections.LEFT ? 0 : null;
+        right = slideFrom === slideFromDirections.RIGHT ? 0 : null;
+        top = `${offset.top}px`;
+        break;
+      case slideFromDirections.BOTTOM:
+        _height = Number.isNaN(Number(height)) ? height : `${height}px`;
+        _width = `calc(100% - ${offset.left}px - ${offset.right}px)`;
+        left = `${offset.left}px`;
+        right = `${offset.right}px`;
+        bottom = 0;
+        borderTop = `1px solid ${tokens.border.color}`;
+        break;
+      default:
+    }
 
     const animationStyle = isAnimating
       ? css`
@@ -62,22 +114,22 @@ export const Dialog = styled.div(
       ${animationStyle};
       background: ${tokens.color.white};
       border: 0;
+      border-top: ${borderTop};
+      bottom: ${bottom};
       box-shadow: ${boxShadow};
       box-sizing: border-box;
       ${childPanel}
       display: flex;
       flex-direction: column;
-      height: 100%;
-      ${offsetY ? `height: calc(100% - ${offsetY}px);` : ""}
-      ${slideFrom === slideFromDirections.LEFT ? `left: 0;` : `right: 0;`}
+      height: ${_height};
+      left: ${left};
       margin: 0;
       overflow: auto;
       padding: 0;
       ${isInline ? "position: relative;" : "position: fixed;"}
-      right: 0;
-      top: 0;
-      top: ${offsetY}px;
-      width: ${widthAsPx};
+      right: ${right};
+      top: ${top};
+      width: ${_width};
       z-index: ${zIndex};
 
       &:focus {