+++ packages/Panel/CHANGELOG.md
@@ -11,6 +11,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 ### Removed
 
 - Breaking change: removed `isSlideFromLeft` prop. Now you must use `slideFrom={Panel.slideFromDirections.LEFT}`.
+- Breaking change: removed `offsetY` prop. Now you must use `offset={{ top: 123 }}`.
 
 ### Added